+++ src/http/multipart.rs
@@ -39,7 +39,7 @@ impl MultipartOptions {
 }
 
 pub(super) async fn receive_batch_multipart(
-    body: impl AsyncRead + Send + 'static,
+    body: impl AsyncRead + Send,
     boundary: impl Into<String>,
     opts: MultipartOptions,
 ) -> Result<BatchRequest, ParseRequestError> {